# docs/api-specification.yaml
openapi: 3.0.0
paths:
...
  /api/v1/auth/login:
    post:
      operationId: AuthController_login_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Logged in — session cookies set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUserDto'
        '401':
          description: Invalid credentials, disabled, or locked account
      summary: Log in with email + password
      tags: &ref_0
        - auth
...
  /api/v1/auth/sessions/{id}:
    delete:
      operationId: AuthController_revokeSession_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Revoke a specific session (must belong to the current user)
      tags: *ref_0
...
info:
  title: PayFlow Orchestration Layer
  description: >-
    Production-grade payment orchestration API routing transactions across
    Razorpay, Stripe, PayU, UPI, Paystack, Flutterwave, Interswitch, and Opay
    with intelligent failover, idempotency, and complete audit trails.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    X-API-Key:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: ops-admin@payflow.example
        password:
          type: string
          minLength: 8
          format: password
      required:
        - email
        - password
    SessionUserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - OPS_ADMIN
            - OPS_VIEWER
            - MERCHANT_ADMIN
            - MERCHANT_VIEWER
        merchantId:
          type: object
          format: uuid
          nullable: true
      required:
        - id
        - email
        - role
        - merchantId
...